my_project/src/taxi_app/requirements.txt
databricks-sdk>=0.20.0
databricks-sql-connector>=3.0.0
plotly>=5.0.0
pandas>=2.0.0
